starting to expose the ts-functionality, first for play and fun, later for replacement of existing ts-exposure
