+++ test/test.cbp
@@ -12,7 +12,7 @@
 				<Option object_output="obj/Debug/" />
 				<Option type="1" />
 				<Option compiler="gcc" />
-				<Option parameters="time_axis_test" />
+				<Option parameters="timeseries_test test_api_ts" />
 				<Compiler>
 					<Add option="-g" />
 				</Compiler>
@@ -61,6 +61,7 @@
 		<Linker>
 			<Add option="-pthread" />
 			<Add library="shyftcore" />
+			<Add library="shyftapi" />
 			<Add library="pthread" />
 			<Add library="boost_filesystem" />
 			<Add library="boost_system" />
